Controller changes

+++ src/main/java/ntrusted/mongoModel/Connection.java
@@ -2,9 +2,9 @@ package ntrusted.mongoModel;
 
 public class Connection {
 	private String userid;
-	private double RenterTrust;
+	private float RenterTrust;
 	private int RenterNoTrx;
-	private double RenteTrust;
+	private float RenteTrust;
 	private int RenteeNoTrx;
 	private int activeFriend;
 	
@@ -14,10 +14,10 @@ public class Connection {
 	public void setUserid(String userid) {
 		this.userid = userid;
 	}
-	public double getRenterTrust() {
+	public float getRenterTrust() {
 		return RenterTrust;
 	}
-	public void setRenterTrust(double renterTrust) {
+	public void setRenterTrust(float renterTrust) {
 		RenterTrust = renterTrust;
 	}
 	public int getRenterNoTrx() {
@@ -26,10 +26,10 @@ public class Connection {
 	public void setRenterNoTrx(int renterNoTrx) {
 		RenterNoTrx = renterNoTrx;
 	}
-	public double getRenteTrust() {
+	public float getRenteTrust() {
 		return RenteTrust;
 	}
-	public void setRenteTrust(double renteTrust) {
+	public void setRenteTrust(float renteTrust) {
 		RenteTrust = renteTrust;
 	}
 	public int getRenteeNoTrx() {
